requirements.txt
asgiref==3.3.1
Django==3.0
django-taggit==1.2.0
Markdown==3.2.1
pytz==2020.5
sqlparse==0.4.1
